add pending status to profiles and users page

+++ back-end/src/users/users.controller.ts
@@ -20,6 +20,7 @@ import { FileInterceptor } from '@nestjs/platform-express';
 import JwtTwoFactorGuard from 'src/auth/jwt-2fa-auth.guard';
 import { saveImageToStorage } from './helpers/image-storage';
 import { BlockGuard } from './block.guard';
+import User from './entities/user.entity';
 
 @Controller('users')
 export class UsersController {
@@ -57,6 +58,11 @@ export class UsersController {
     @Req() req: RequestWithUser,
     @Body() body: { relatedUserId: string },
   ) {
+    if (req.user.id === body.relatedUserId)
+      throw new HttpException(
+        'Forbidden, Can not send request',
+        HttpStatus.FORBIDDEN,
+      );
     return await this.usersService.sendRequest(req.user.id, body.relatedUserId);
   }
 
@@ -66,6 +72,11 @@ export class UsersController {
     @Req() req: RequestWithUser,
     @Body() body: { relatedUserId: string },
   ) {
+    if (req.user.id === body.relatedUserId)
+      throw new HttpException(
+        'Forbidden, Can not send request',
+        HttpStatus.FORBIDDEN,
+      );
     return await this.usersService.acceptRequest(
       req.user.id,
       body.relatedUserId,
@@ -78,6 +89,11 @@ export class UsersController {
     @Req() req: RequestWithUser,
     @Body() body: { relatedUserId: string },
   ) {
+    if (req.user.id === body.relatedUserId)
+      throw new HttpException(
+        'Forbidden, Can not send request',
+        HttpStatus.FORBIDDEN,
+      );
     return await this.usersService.removeRelation(
       req.user.id,
       body.relatedUserId,
@@ -90,6 +106,11 @@ export class UsersController {
     @Req() req: RequestWithUser,
     @Body() body: { relatedUserId: string },
   ) {
+    if (req.user.id === body.relatedUserId)
+      throw new HttpException(
+        'Forbidden, Can not send request',
+        HttpStatus.FORBIDDEN,
+      );
     return await this.usersService.blockUser(req.user.id, body.relatedUserId);
   }
 
@@ -99,12 +120,18 @@ export class UsersController {
     @Req() req: RequestWithUser,
     @Body() body: { relatedUserId: string },
   ) {
+    if (req.user.id === body.relatedUserId)
+      throw new HttpException(
+        'Forbidden, Can not send request',
+        HttpStatus.FORBIDDEN,
+      );
     return await this.usersService.unblockUser(req.user.id, body.relatedUserId);
   }
 
   @Get('id/:id/friends')
-  async getFriends(@Param('id') id: string) {
-    return await this.usersService.getFriends(id);
+  @UseGuards(JwtTwoFactorGuard)
+  async getFriends(@Param('id') userId: string) {
+    return await this.usersService.getFriends(userId);
   }
 
   @Get('blocked')
